+++ src/server/database/default/datasource.ts
@@ -5,13 +5,12 @@ import { DataSource } from 'typeorm';
 export * from './entity';
 
 import { User } from './entity/User';
-import { Session } from './entity/Session';
 import { Role } from './entity/Role';
 import { Replicant } from './entity/Replicant';
 import { Permission } from './entity/Permission';
 import { Identity } from './entity/Identity';
 import { ApiKey } from './entity/ApiKey';
-import { nodecgRootPath } from '../../shared/utils/rootPath';
+import { nodecgRootPath } from '../../../shared/utils/rootPath';
 
 const dbPath = path.join(nodecgRootPath, 'db/nodecg.sqlite3');
 export const testing = process.env.NODECG_TEST?.toLowerCase() === 'true';
@@ -32,8 +31,8 @@ const dataSource = new DataSource({
 	 */
 	database: testing ? ':memory:' : dbPath,
 	logging: false,
-	entities: [ApiKey, Identity, Permission, Replicant, Role, Session, User],
-	migrations: [path.join(nodecgRootPath, 'out/server/database/migration/*.js')],
+	entities: [ApiKey, Identity, Permission, Replicant, Role, User],
+	migrations: [path.join(nodecgRootPath, 'out/server/database/default/migration/*.js')],
 	migrationsRun: true,
 	synchronize: false,
 });