+++ src/server/database/default/entity/ApiKey.ts
@@ -1,8 +1,9 @@
 import { Entity, ManyToOne, Generated, PrimaryColumn } from 'typeorm';
 import { User } from './User';
+import type { ApiKey as ApiKeyModel } from '../../models';
 
 @Entity()
-export class ApiKey {
+export class ApiKey implements ApiKeyModel {
 	@PrimaryColumn()
 	@Generated('uuid')
 	secret_key!: string;