+++ src/server/server/index.ts
@@ -48,7 +48,6 @@ import createLogger from '../logger';
 import socketAuthMiddleware from '../login/socketAuthMiddleware';
 import socketApiMiddleware from './socketApiMiddleware';
 import Replicator from '../replicant/replicator';
-import * as db from '../database';
 import type { ClientToServerEvents, ServerToClientEvents, TypedSocketServer } from '../../types/socket-protocol';
 import GraphicsLib from '../graphics';
 import { DashboardLib } from '../dashboard';
@@ -62,6 +61,7 @@ import type { NodeCG } from '../../types/nodecg';
 import { TypedEmitter } from '../../shared/typed-emitter';
 import { nodecgRootPath } from '../../shared/utils/rootPath';
 import { NODECG_ROOT } from '../nodecg-root';
+import { getAllReplicants } from '../database/default/utils';
 
 const renderTemplate = memoize((content, options) => template(content)(options));
 
@@ -140,7 +140,6 @@ export class NodeCGServer extends TypedEmitter<EventMap> {
 		const io = this._io.of('/');
 		log.info('Starting NodeCG %s (Running on Node.js %s)', pjson.version, process.version);
 
-		const database = await db.getConnection();
 		if (sentryEnabled) {
 			app.use(Sentry.Handlers.requestHandler());
 		}
@@ -165,7 +164,7 @@ export class NodeCGServer extends TypedEmitter<EventMap> {
 		if (config.login?.enabled) {
 			log.info('Login security enabled');
 			const login = await import('../login');
-			const { app: loginMiddleware, sessionMiddleware } = await login.createMiddleware({
+			const { app: loginMiddleware, sessionMiddleware } = login.createMiddleware({
 				onLogin: (user) => {
 					// If the user had no roles, then that means they "logged in"
 					// with a third-party auth provider but weren't able to
@@ -254,7 +253,7 @@ export class NodeCGServer extends TypedEmitter<EventMap> {
 			app.use(sentryHelpers.app);
 		}
 
-		const persistedReplicantEntities = await database.getRepository(db.Replicant).find();
+		const persistedReplicantEntities = await getAllReplicants();
 		const replicator = new Replicator(io, persistedReplicantEntities);
 		this._replicator = replicator;