+++ test/core.ts
@@ -1,6 +1,5 @@
 // Packages
 import test from 'ava';
-import fetch from 'node-fetch-commonjs';
 
 // Ours
 import * as server from './helpers/server';